fix: move /authors to root path instead of /feeds/authors
- Move /authors endpoint from FeedsController to AppController
- Keep /feeds/* endpoints under /feeds path
