+++ apps/server/src/app.controller.ts
@@ -3,17 +3,45 @@ import { AppService } from './app.service';
 import { ConfigService } from '@nestjs/config';
 import { ConfigurationType } from './configuration';
 import { Response as Res } from 'express';
+import { PrismaService } from '@server/prisma/prisma.service';
 
 @Controller()
 export class AppController {
   constructor(
     private readonly appService: AppService,
     private readonly configService: ConfigService,
+    private readonly prismaService: PrismaService,
   ) {}
 
   @Get()
   getHello(): string {
     return this.appService.getHello();
+  }
+
+  @Get('/authors')
+  async getAuthors() {
+    const feeds = await this.prismaService.feed.findMany({
+      select: {
+        id: true,
+        mpName: true,
+        mpIntro: true,
+        mpCover: true,
+        status: true,
+        syncTime: true,
+        updateTime: true,
+      },
+    });
+    return {
+      items: feeds.map((f) => ({
+        mp_id: f.id,
+        author_name: f.mpName,
+        intro: f.mpIntro,
+        cover: f.mpCover,
+        status: f.status,
+        sync_time: f.syncTime,
+        update_time: f.updateTime,
+      })),
+    };
   }
 
   @Get('/robots.txt')
